Git Reposu Güncellendi

+++ CSharp/NetMVC/HalkOdePaymentIntegration/Models/Contract/Response/GetPosResponse.cs
@@ -14,11 +14,11 @@ namespace HalkOdePaymentIntegration.Contract.Response
         public int pos_id { get; set; }
         public int campaign_id { get; set; }
         public int allocation_id { get; set; }
-        public int installments_number { get; set; }
+        public double installments_number { get; set; }
         public string card_type { get; set; }
         public string card_program { get; set; }
         public string card_scheme { get; set; }
-        public int payable_amount { get; set; }
+        public double payable_amount { get; set; }
         public string hash_key { get; set; }
         public string amount_to_be_paid { get; set; }
         public string currency_code { get; set; }
